requirements.txt
python-telegram-bot==20.8
openai==1.30.1
gTTS==2.5.1
